test(core): add MyNodeLLTest.java and MyLinkedListTest.java
Node test class is complete and covers the necessary functions and
equivalence partitions. The LinkedListTest.java contains equivalence
partition tests for adding 1 element (front of linked list) and adding
2 elements (front and back of list) and lastly, adding 3 elements tested
and passed. Linked list tests cover all the necessary functions in the
tests including branch coverage. Hence, coverage so far is 100%, only
thing not covered are the getters and setters which are not very
important as of yet.